requirements.txt
certifi==2022.12.7
charset-normalizer==2.1.1
idna==3.4
prometheus-client==0.16.0
requests==2.28.1
urllib3==1.26.13
python-dotenv==0.21.1
